Fix some logic placement

+++ src/main/java/com/mycompany/coursemanagement/BUS/OnsiteCourseBUS.java
@@ -5,7 +5,9 @@
 package com.mycompany.coursemanagement.BUS;
 
 import com.mycompany.coursemanagement.DAO.CourseDAO;
+import com.mycompany.coursemanagement.DAO.CourseInstructorDAO;
 import com.mycompany.coursemanagement.DAO.OnsiteCourseDAO;
+import com.mycompany.coursemanagement.DAO.StudentGradeDAO;
 import com.mycompany.coursemanagement.Models.Course;
 import com.mycompany.coursemanagement.Models.Department;
 import com.mycompany.coursemanagement.Models.OnsiteCourse;
@@ -22,6 +24,8 @@ public class OnsiteCourseBUS {
 
     private final OnsiteCourseDAO ocDAO = new OnsiteCourseDAO();
     private final CourseDAO courseDAO = new CourseDAO();
+    private final CourseInstructorDAO courseInstructorDAO = new CourseInstructorDAO();
+    private final StudentGradeDAO studentGradeDAO = new StudentGradeDAO();
 
     public OnsiteCourseBUS() {
     }
@@ -126,7 +130,21 @@ public class OnsiteCourseBUS {
         return result;
     }
 
-    public int deleteOnsiteCourse(int courseId) throws SQLException {
+    public int deleteOnsiteCourse(int courseId, String courseTitle) throws SQLException, IllegalArgumentException {
+        boolean canBeDeleted = true;
+        try {
+            canBeDeleted = !courseInstructorDAO.anyInstructorFound(courseId) && !studentGradeDAO.anyEnrollmentFound(courseId);
+        } catch (SQLException e) {
+            throw e;
+        } catch (Exception e) {
+            throw e;
+        }
+
+        if (!canBeDeleted) {
+            throw new IllegalArgumentException("Course '" + courseTitle + "' (" + courseId + ") cannot be deleted as it already has instructors assigned or students enrolled.");
+        }
+
         return ocDAO.deleteOnsiteCourse(courseId);
     }
+
 }
